# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for NutriApp - Nutritionist management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create patient (also creates login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Patient created with credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientCreated"
        "409":
          description: Email already in use
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDetail"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /patients/{id}/reset-password:
    post:
      operationId: resetPatientPassword
      tags: [patients]
      summary: Reset patient password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetResult"

  # ── Groups ─────────────────────────────────────────────
...
  /portal/feed:
    get:
      operationId: getPatientFeed
      tags: [portal]
      summary: Get patient's content feed
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, ebook, receita, video, artigo]
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Publication"
...
  /portal/consultations/{consultationId}:
    get:
      operationId: getMyConsultation
      tags: [portal]
      summary: Get patient's consultation detail
      parameters:
        - name: consultationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, name, email, createdAt, hasAccess]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        hasAccess:
          type: boolean
        createdAt:
          type: string
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
    PatientDetail:
      type: object
      required: [id, name, email, createdAt, hasAccess, groups, consultationCount]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        hasAccess:
          type: boolean
        createdAt:
          type: string
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
        consultationCount:
          type: integer
    PatientCreated:
      type: object
      required: [patient, provisionalPassword]
      properties:
        patient:
          $ref: "#/components/schemas/Patient"
        provisionalPassword:
          type: string
    PatientInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        phone:
          type: string
        birthDate:
          type: string
        sex:
          type: string
          enum: [masculino, feminino, outro]
        groupIds:
          type: array
          items:
            type: integer
    PatientUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        sex:
          type: ["string", "null"]
        groupIds:
          type: array
          items:
            type: integer
    ResetPasswordInput:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [generate, manual]
        newPassword:
          type: ["string", "null"]
    PasswordResetResult:
      type: object
      required: [message]
      properties:
        message:
          type: string
        provisionalPassword:
          type: ["string", "null"]

    # ── Group Schemas ─────────────────────────────────────
    GroupSummary:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
...
    ConsultationDetail:
      type: object
      required: [id, patientId, type, date, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        type:
          type: string
          enum: [avaliacao_inicial, cardapio, reavaliacao]
        date:
          type: string
        createdAt:
          type: string
        # Avaliação Inicial
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        bmi:
          type: ["number", "null"]
        waistCm:
          type: ["number", "null"]
        hipCm:
          type: ["number", "null"]
        abdomenCm:
          type: ["number", "null"]
        armCm:
          type: ["number", "null"]
        thighCm:
          type: ["number", "null"]
        calfCm:
          type: ["number", "null"]
        bioimpedancePdfPath:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        # Cardápio
        objective:
          type: ["string", "null"]
        vetKcal:
          type: ["number", "null"]
        menuPdfPath:
          type: ["string", "null"]
        restrictions:
          type: ["string", "null"]
        additionalGuidance:
          type: ["string", "null"]
        # Reavaliação
        baselineConsultationId:
          type: ["integer", "null"]
        evolutionNotes:
          type: ["string", "null"]
        comparison:
          $ref: "#/components/schemas/ConsultationComparison"
    ConsultationComparison:
      type: object
      properties:
        weightDelta:
          type: ["number", "null"]
        bmiDelta:
          type: ["number", "null"]
        waistDelta:
          type: ["number", "null"]
        hipDelta:
          type: ["number", "null"]
        abdomenDelta:
          type: ["number", "null"]
        armDelta:
          type: ["number", "null"]
        thighDelta:
          type: ["number", "null"]
        calfDelta:
          type: ["number", "null"]
...
    Publication:
      type: object
      required: [id, title, category, status, visibility, createdAt, publishedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [ebook, receita, video, artigo]
        status:
          type: string
          enum: [rascunho, publicado]
        visibility:
          type: string
          enum: [geral, grupos, pacientes]
        videoUrl:
          type: ["string", "null"]
        pdfPath:
          type: ["string", "null"]
        imagePath:
          type: ["string", "null"]
        groupIds:
          type: array
          items:
            type: integer
        patientIds:
          type: array
          items:
            type: integer
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
        createdAt:
          type: string
        publishedAt:
          type: ["string", "null"]
        isNew:
          type: boolean
...